Controlling the Number of Instances
With this change, Symphony now can only have one instance of the object
created. This is to prevent the errors that occur with trying to open
multiple instances.

to instansiate the instance the code is:
symphonyInstance = Symphony.getInstance;

to gain access to that instance (for example to delete the object) you
just use the above code and the instance will be retrieved.