Multi-Channel Functionality
This version of Symphony changes the way the parameters are stored and
accessed throught the different class files SymphonyProtocol, the
underlying protocols and editParameters.

This version has not yet been fully tested. Therefore there are still
some minor bugs in the application and a couple of features not yet
implemented.

The continuation from these changes will provide the ability to log the
experimental changes (ie... changing solution) within the UI parameter
and attach it to the epoch.

In addition there will be the functionality to have default values
stored for each channel/LED (through a config file or ?????)